Move code from Home-index.js to expensesContext
